Fix a potential bug with context configs.
This issue is related to how strings for select backends are not deeply
copied, but rather only have their pointers copied. This results in a
situation where the client may delete the strings they passed to the
context config assuming miniaudio makes a copy of it when in fact it
hasn't.