Improvements to high order low- and high-pass filters.
This changes the Q value of the chain of second order low-pass filters
making up the high order filters. Previously the Q value was always set
to 0.707107 for the entire chain, but this is incorrect for Butterworth
filters when the order is greater than 2.

This should make a small improvement to the quality of the linear
resampler.