Set up some infrastructure for transitioning to the new callbacks.
With the introduction of custom backends, it became necessary to
simplify the backend callbacks. For backwards compatibility, all
built-in backends are still using the old callbacks. Future commits are
going to be transitioning built-in backends to the new system. This
commit is designed to help with this migration so that we can migrate
backends one at a time while still keeping compatibility with this old
system, thereby allowing backends that haven't yet transitioned to keep
working.