Fix use of uninitialized variables
When shutting down, the resultALSA variable can be used when
uninitialized in the ma_device_read__alsa and ma_device_write__alsa
routines.  This has been verified by both gcc warnings and in
application code (we were occasionally seeing segfaults when our
callback routine was being passed a crazy large value for the number of
frames read when shutting down).

Signed-off-by: Charles Steinkuehler <charles@steinkuehler.net>