Introduce new safe atomic types.
The idea behind this is to avoid subtle bugs that are a result of
accidentally not using atomic operations to reference variables that
must always be used atomically.

This commit does not convert every atomic variable to the new system,
however going forward this will need to be adopted over time.