Fix a bug with asynchronously loaded sounds and ranges and looping.
The issue with this one is that it was possible for the asynchronous
loading process to override settings that were explicitly set via
ma_data_source_set_*(). The fix for this is to only set these on the
asynchronous side if the values in the resource manager's data source
config is not equal to defaults.