Get playback working on Windows 98 with the DirectSound backend.
This change adds a fallback for when setting the format of the primary
buffer fails. I have observed this happening with floating point
formats so I've forced the use of integer PCM formats in the fallback
case. I've also restricted it to 44100 because when I tried 4800 on
my Windows 98 machine it resulted in incorrect output.