Fix crash when running on Windows 95.
It looks like CoInitializeEx() is not available on Windows 95. This
commit changes it so that if CoInitializeEx() cannot be found, such as
the case with Windows 95, it falls back to CoInitialize().

As of this commit, a crash still occurs when trying to initialize the
DirectSound backend. The WinMM backend works, however.